requirements.txt
python-telegram-bot==21.7
SpeechRecognition==3.10.4
pydub==0.25.1
gspread==6.2.1
google-auth==2.49.0
